refactor(transport): extract DDS subscriber path into its own file
Moves createSubscriber, appendSubscriber, takeAllSubscribers, and the
DDSTransportSubscriberImpl concrete type into a focused extension file.
DDSTransportSubscriberImpl stays default-internal (not private) because
the core file's 'subscribers' array references it by name.

Demotes a minimum-necessary set of DDSTransportSession members from
private to default-internal so the cross-file extension compiles:
client, publishers, subscribers, lock, _isOpen, bridgeQoS. None of
these are public today and none become public; the change is invisible
at the module boundary. Verified with swift package
diagnose-api-breaking-changes 0.6.0.
